Add support for 'async' reload in 'network' role.
When the 'network' role is used to change the address of the network
that Ansible is using to manage the system, a synchronous reload will
hang forever while Ansible tries to connect to the old address. Using
an 'async' reload allows Ansible to rediscover the system's address
and make a new connection.

Closes #51.